requirements.txt
certifi==2020.6.20
cffi==1.14.1
cryptography==3.0
decorator==4.4.2
pycparser==2.20
python-dotenv==0.14.0
python-telegram-bot[job-queue]==22.5
six==1.15.0
tornado==6.5.2
pytz==2025.2
